requirements.txt
pytest
allure-pytest
fastapi
uvicorn
requests
flask
python-multipart